allow --check-tmpdir arg to be configurable
Addresses: https://github.com/redhat-openstack/tripleo-quickstart/issues/66

virt-sparsify checks to see if there is enough space using an estimate based on the
virtual size of the disk (40G in our case). The actual size is substantially smaller.
The default behavior for virt-sparsify is "warn" - which in this context will cause
the shell task to hang forever waiting for a CR to ack the warning.  We presently use
"fail" which is the safest option and will never cause a nasty failure due to a volume
that is completely filled up.  This change permits a pragmatic middle ground, where
we still default to the safe option (fail) but allow for pragmatism (I have 20G free
but not 40....do it anyway).

Change-Id: I78263f6ac06f3184afb3ae7339b2ee3f40d282ec